Work on interpolation types
