+++ cpp/dolfinx/mesh/utils.cpp
@@ -31,46 +31,46 @@ using cmdspan3x_t
 namespace
 {
 
-/// The coordinates for all 'vertices' in the mesh
-/// @param[in] mesh The mesh to compute the vertex coordinates for
-/// @return The vertex coordinates. The shape is `(3, num_vertices)` and
-/// the jth column hold the coordinates of vertex j.
-std::pair<std::vector<double>, std::array<std::size_t, 2>>
-compute_vertex_coords(const mesh::Mesh<double>& mesh)
-{
-  const mesh::Topology& topology = mesh.topology();
-  const int tdim = topology.dim();
+// /// The coordinates for all 'vertices' in the mesh
+// /// @param[in] mesh The mesh to compute the vertex coordinates for
+// /// @return The vertex coordinates. The shape is `(3, num_vertices)` and
+// /// the jth column hold the coordinates of vertex j.
+// std::pair<std::vector<double>, std::array<std::size_t, 2>>
+// compute_vertex_coords(const mesh::Mesh<double>& mesh)
+// {
+//   const mesh::Topology& topology = mesh.topology();
+//   const int tdim = topology.dim();
 
-  // Create entities and connectivities
-  mesh.topology_mutable().create_connectivity(tdim, 0);
+//   // Create entities and connectivities
+//   mesh.topology_mutable().create_connectivity(tdim, 0);
 
-  // Get all vertex 'node' indices
-  const graph::AdjacencyList<std::int32_t>& x_dofmap = mesh.geometry().dofmap();
-  const std::int32_t num_vertices = topology.index_map(0)->size_local()
-                                    + topology.index_map(0)->num_ghosts();
-  auto c_to_v = topology.connectivity(tdim, 0);
-  assert(c_to_v);
-  std::vector<std::int32_t> vertex_to_node(num_vertices);
-  for (int c = 0; c < c_to_v->num_nodes(); ++c)
-  {
-    auto x_dofs = x_dofmap.links(c);
-    auto vertices = c_to_v->links(c);
-    for (std::size_t i = 0; i < vertices.size(); ++i)
-      vertex_to_node[vertices[i]] = x_dofs[i];
-  }
+//   // Get all vertex 'node' indices
+//   const graph::AdjacencyList<std::int32_t>& x_dofmap = mesh.geometry().dofmap();
+//   const std::int32_t num_vertices = topology.index_map(0)->size_local()
+//                                     + topology.index_map(0)->num_ghosts();
+//   auto c_to_v = topology.connectivity(tdim, 0);
+//   assert(c_to_v);
+//   std::vector<std::int32_t> vertex_to_node(num_vertices);
+//   for (int c = 0; c < c_to_v->num_nodes(); ++c)
+//   {
+//     auto x_dofs = x_dofmap.links(c);
+//     auto vertices = c_to_v->links(c);
+//     for (std::size_t i = 0; i < vertices.size(); ++i)
+//       vertex_to_node[vertices[i]] = x_dofs[i];
+//   }
 
-  // Pack coordinates of vertices
-  std::span<const double> x_nodes = mesh.geometry().x();
-  std::vector<double> x_vertices(3 * vertex_to_node.size(), 0.0);
-  for (std::size_t i = 0; i < vertex_to_node.size(); ++i)
-  {
-    const int pos = 3 * vertex_to_node[i];
-    for (std::size_t j = 0; j < 3; ++j)
-      x_vertices[j * vertex_to_node.size() + i] = x_nodes[pos + j];
-  }
+//   // Pack coordinates of vertices
+//   std::span<const double> x_nodes = mesh.geometry().x();
+//   std::vector<double> x_vertices(3 * vertex_to_node.size(), 0.0);
+//   for (std::size_t i = 0; i < vertex_to_node.size(); ++i)
+//   {
+//     const int pos = 3 * vertex_to_node[i];
+//     for (std::size_t j = 0; j < 3; ++j)
+//       x_vertices[j * vertex_to_node.size() + i] = x_nodes[pos + j];
+//   }
 
-  return {std::move(x_vertices), {3, vertex_to_node.size()}};
-}
+//   return {std::move(x_vertices), {3, vertex_to_node.size()}};
+// }
 
 /// The coordinates of 'vertices' for for entities of a give dimension
 /// that are attached to specified facets.
@@ -385,53 +385,53 @@ mesh::compute_midpoints(const Mesh<double>& mesh, int dim,
   return x_mid;
 }
 //-----------------------------------------------------------------------------
-std::vector<std::int32_t> mesh::locate_entities(
-    const Mesh<double>& mesh, int dim,
-    const std::function<std::vector<std::int8_t>(
-        std::experimental::mdspan<
-            const double,
-            std::experimental::extents<
-                std::size_t, 3, std::experimental::dynamic_extent>>)>& marker)
-{
-  // Run marker function on vertex coordinates
-  const auto [xdata, xshape] = compute_vertex_coords(mesh);
-  cmdspan3x_t x(xdata.data(), xshape);
-  const std::vector<std::int8_t> marked = marker(x);
-  if (marked.size() != x.extent(1))
-    throw std::runtime_error("Length of array of markers is wrong.");
+// std::vector<std::int32_t> mesh::locate_entities(
+//     const Mesh<double>& mesh, int dim,
+//     const std::function<std::vector<std::int8_t>(
+//         std::experimental::mdspan<
+//             const double,
+//             std::experimental::extents<
+//                 std::size_t, 3, std::experimental::dynamic_extent>>)>& marker)
+// {
+//   // Run marker function on vertex coordinates
+//   const auto [xdata, xshape] = compute_vertex_coords(mesh);
+//   cmdspan3x_t x(xdata.data(), xshape);
+//   const std::vector<std::int8_t> marked = marker(x);
+//   if (marked.size() != x.extent(1))
+//     throw std::runtime_error("Length of array of markers is wrong.");
 
-  const mesh::Topology& topology = mesh.topology();
-  const int tdim = topology.dim();
+//   const mesh::Topology& topology = mesh.topology();
+//   const int tdim = topology.dim();
 
-  mesh.topology_mutable().create_entities(dim);
-  mesh.topology_mutable().create_connectivity(tdim, 0);
-  if (dim < tdim)
-    mesh.topology_mutable().create_connectivity(dim, 0);
+//   mesh.topology_mutable().create_entities(dim);
+//   mesh.topology_mutable().create_connectivity(tdim, 0);
+//   if (dim < tdim)
+//     mesh.topology_mutable().create_connectivity(dim, 0);
 
-  // Iterate over entities of dimension 'dim' to build vector of marked
-  // entities
-  auto e_to_v = topology.connectivity(dim, 0);
-  assert(e_to_v);
-  std::vector<std::int32_t> entities;
-  for (int e = 0; e < e_to_v->num_nodes(); ++e)
-  {
-    // Iterate over entity vertices
-    bool all_vertices_marked = true;
-    for (std::int32_t v : e_to_v->links(e))
-    {
-      if (!marked[v])
-      {
-        all_vertices_marked = false;
-        break;
-      }
-    }
+//   // Iterate over entities of dimension 'dim' to build vector of marked
+//   // entities
+//   auto e_to_v = topology.connectivity(dim, 0);
+//   assert(e_to_v);
+//   std::vector<std::int32_t> entities;
+//   for (int e = 0; e < e_to_v->num_nodes(); ++e)
+//   {
+//     // Iterate over entity vertices
+//     bool all_vertices_marked = true;
+//     for (std::int32_t v : e_to_v->links(e))
+//     {
+//       if (!marked[v])
+//       {
+//         all_vertices_marked = false;
+//         break;
+//       }
+//     }
 
-    if (all_vertices_marked)
-      entities.push_back(e);
-  }
+//     if (all_vertices_marked)
+//       entities.push_back(e);
+//   }
 
-  return entities;
-}
+//   return entities;
+// }
 //-----------------------------------------------------------------------------
 std::vector<std::int32_t> mesh::locate_entities_boundary(
     const Mesh<double>& mesh, int dim,
